moved to mvvm architechture
