+++ lib/1view/pages/favourite page/favourites_page.dart
@@ -1,7 +1,8 @@
-import '../../logic/favourite_breeds.dart';
+import 'package:dog_app/1view/widgets/main_header.dart';
+import 'package:dog_app/1view/widgets/main_list.dart';
+
+import '../../../2viewmodel/favourite_breeds_viewmodel.dart';
 import 'package:provider/provider.dart';
-import '../widgets/main_header.dart';
-import '../widgets/main_list.dart';
 import 'package:flutter/material.dart';
