+++ lib/1view/pages/home detail page/home_detail_page.dart
@@ -1,7 +1,7 @@
-import '../../logic/favourite_breeds.dart';
+import 'package:dog_app/3model/models/breeds.dart';
+import '../../../2viewmodel/favourite_breeds_viewmodel.dart';
 import 'package:provider/provider.dart';
-import '../../data/models/breeds.dart';
-import '../widgets/description.dart';
+import 'widgets/description.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:cached_network_image/cached_network_image.dart';
