+++ lib/1view/pages/home detail page/widgets/description.dart
@@ -1,4 +1,4 @@
-import '../../data/models/breeds.dart';
+import 'package:dog_app/3model/models/breeds.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
